divide into two routes
