+++ src/router/index.js
@@ -1,6 +1,7 @@
 import Vue from 'vue';
 import Router from 'vue-router';
 import editPage from '../components/EditPage';
+import newFilePage from '../components/NewFilePage';
 
 Vue.use(Router);
 
@@ -8,6 +9,8 @@ Vue.use(Router);
 export default new Router({
   routes: [{
     path: '/',
+    name: 'newFilePage',
+    component: newFilePage,
   },
   {
     path: '/editPage',
